# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Interactive Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /series:
    get:
      operationId: listSeries
      tags: [series]
      summary: List all book series
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Series"
    post:
      operationId: createSeries
      tags: [series]
      summary: Create a new series (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeriesRequest"
      responses:
        "201":
          description: Series created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /series/{seriesId}:
    get:
      operationId: getSeries
      tags: [series]
      summary: Get a series with its classes
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Series detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesWithClasses"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSeries
      tags: [series]
      summary: Update a series (admin)
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSeriesRequest"
      responses:
        "200":
          description: Series updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSeries
      tags: [series]
      summary: Delete a series (admin)
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /series/{seriesId}/classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List classes in a series
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class level (admin)
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Series:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        classCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, classCount, createdAt]
    SeriesWithClasses:
      allOf:
        - $ref: "#/components/schemas/Series"
        - type: object
          properties:
            classes:
              type: array
              items:
                $ref: "#/components/schemas/Class"
          required: [classes]
    CreateSeriesRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
      required: [name, slug]
    UpdateSeriesRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
    Class:
      type: object
      properties:
        id:
          type: integer
        seriesId:
          type: integer
        classLevel:
          type: string
        label:
          type: string
        sortOrder:
          type: integer
        contentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, seriesId, classLevel, label, sortOrder, contentCount, createdAt]
...
    CreateClassRequest:
      type: object
      properties:
        classLevel:
          type: string
          minLength: 1
        label:
          type: string
          minLength: 1
        sortOrder:
          type: integer
      required: [classLevel, label]
...